fix(tests): resolve compiler warnings in test files
- bit_crusher_rng_bias_test.cpp: Use 10.0f instead of 10.0 for float division
- lfo_test.cpp: Cast discarded [[nodiscard]] return values to (void)
- bit_crusher_test.cpp: Cast discarded [[nodiscard]] return value to (void)
- tap_manager_test.cpp: Use size_t instead of float for getActiveTapCount()

+++ tests/unit/systems/tap_manager_test.cpp
@@ -730,7 +730,7 @@ TEST_CASE("TapManager: setTempo() updates tempo-synced taps", "[tap-manager][tem
 
     SECTION("Invalid tempo (0 or negative) is ignored") {
         tm.setTempo(120.0f);
-        const float prevActive = tm.getActiveTapCount();
+        const size_t prevActive = tm.getActiveTapCount();
 
         tm.setTempo(0.0f);   // Should be ignored
         tm.setTempo(-100.0f); // Should be ignored
